backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
pypdf
python-dotenv
sqlalchemy
psycopg2-binary
langchain
langchain-community
langchain-text-splitters
vecs
huggingface-hub
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
groq
supabase
i18next
